+++ src/main/java/roomescape/infra/JdbcThemeSlotRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.infra;
 
 import org.springframework.jdbc.core.BatchPreparedStatementSetter;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Repository;
 import roomescape.domain.Theme;
 import roomescape.domain.ThemeSlot;
 import roomescape.domain.Time;
+import roomescape.repository.ThemeSlotRepository;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
