+++ src/main/java/roomescape/infra/JdbcTimeRepository.java
@@ -1,10 +1,11 @@
-package roomescape.repository;
+package roomescape.infra;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Time;
+import roomescape.repository.TimeRepository;
 
 import java.time.LocalTime;
 import java.util.List;
