+++ src/test/java/roomescape/repository/JdbcTimeRepositoryTest.java
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.domain.Time;
-import roomescape.repository.JdbcTimeRepository;
+import roomescape.infra.JdbcTimeRepository;
 
 @JdbcTest
 class JdbcTimeRepositoryTest {
